% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_web_file}
\alias{get_web_file}
\title{get_web_file()}
\usage{
get_web_file(url, name, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{url}{url to file to download}

\item{name}{name for the file being downloaded, e.g. 'my_interesting_data"}

\item{overwrite}{whether or not to overwrite the file}

\item{verbose}{if FALSE, suppress status messages and the progress bar.}
}
\value{
a data_file object, or if failed a filepath to the data
}
\description{
Download a web based file into this packages downloads directory (inst/extdata), or elsewhere is
this has been set with set_download_dirpath()
}
