% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_download_dirpath}
\alias{set_download_dirpath}
\title{set_download_dirpath}
\usage{
set_download_dirpath(dir_path, verbose = TRUE)
}
\arguments{
\item{dir_path}{the directory path to set the download path to}

\item{verbose}{where to print out}
}
\value{
nothing, sets the package's config file as side effect
}
\description{
set_download_dirpath
}
